Move morphology file around
